Revert "Page controller refactoring (#584)"
This reverts commit b278feed2468dbaab69ce91766cc78d792ce6b9f.
