+++ server/package.json
@@ -11,7 +11,7 @@
     "prebuild": "rm -rf .build/*",
     "prestart": "npm run build",
     "start": "AWS_PROFILE=peblio ./scripts/run_with_credstash.sh nodemon .build/server.js",
-    "test": "./node_modules/mocha/bin/mocha --exit --compilers js:babel-core/register --require babel-polyfill test/**/*.js"
+    "test": "./node_modules/mocha/bin/mocha --exit --compilers js:babel-core/register test/**/*.js"
   },
   "author": "",
   "license": "ISC",
