+++ server/src/controllers/pageController.js
@@ -1,10 +1,20 @@
-import { getPage, getExamples } from './pageControllerNew';
 const express = require('express');
+
 const Page = require('../models/page.js');
 const User = require('../models/user.js');
 const Folder = require('../models/folder.js');
 
-export async function savePageAsGuest(req, res) {
+async function getPage(req, res) {
+  Page.find({ id: req.params.pageId }, (err, data) => {
+    if (err) {
+      res.send(err);
+    } else {
+      res.send(data);
+    }
+  });
+}
+
+async function savePageAsGuest(req, res) {
   try {
     const hydratedUser = await User.findOne({ name: 'peblioguest' }).exec();
 
@@ -16,7 +26,7 @@ export async function savePageAsGuest(req, res) {
   }
 }
 
-export async function savePage(req, res) {
+async function savePage(req, res) {
   const user = req.user;
   if (!user) {
     return res.status(403).send({ error: 'Please log in first' });
@@ -33,7 +43,7 @@ export async function savePage(req, res) {
   }
 }
 
-export async function deletePage(req, res) {
+async function deletePage(req, res) {
   const { pageId } = req.params;
   try {
     await Page.deleteOne({ _id: pageId });
@@ -43,7 +53,7 @@ export async function deletePage(req, res) {
   }
 }
 
-export async function updatePage(req, res) {
+function updatePage(req, res) {
   Page.update({ id: req.body.id }, {
     heading: req.body.heading,
     title: req.body.title,
@@ -52,16 +62,16 @@ export async function updatePage(req, res) {
     layout: req.body.layout,
     workspace: req.body.workspace
   },
-    (err, data) => {
-      if (err) {
-        res.send(err);
-      } else {
-        res.send({ data: 'Record has been Inserted..!!' });
-      }
-    });
+  (err, data) => {
+    if (err) {
+      res.send(err);
+    } else {
+      res.send({ data: 'Record has been Inserted..!!' });
+    }
+  });
 }
 
-export async function movePage(req, res) {
+async function movePage(req, res) {
   const user = req.user;
   if (!user) {
     return res.status(403).send({ error: 'Please log in first' });
@@ -86,7 +96,7 @@ export async function movePage(req, res) {
         return res.status(404).send({ error: `Folder with id ${folderId} not found` });
       }
       page.folder = folderId;
-      // otherwise, move the page to the top level (remove its folder ID)
+    // otherwise, move the page to the top level (remove its folder ID)
     } else {
       page.folder = undefined;
       // could not use delete page.folder -
